+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/ui/settings/GlobsConfigurable.kt
@@ -37,6 +37,7 @@ class GlobsConfigurable : BoundConfigurable(message("settings.globs.group.title"
                     .disableDownAction()
                     .createPanel()
             ).align(Align.FILL)
+                .comment(message("settings.globs.comment"))
         }.resizableRow()
         row {
             reportBugLink()